requirements.txt
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
odfpy==1.4.1
fpdf2==2.7.8
matplotlib==3.8.3
